target/i386: FXSAVE's template said the instruction cannot touch memory
`48 0f ae 40 40` is fxsave64 0x40(%rax).  It moves 512 bytes and the
wire said it moves none: max_dep_loads and max_dep_stores both 0, which
is the one condition cst_lint's memop-impossible rule exists to flag.
Measured on an x86_64 system trace, 990 impossible memop attributions
landed on that one encoding -- after two of the three encodings in the
class were closed at gen_movl_seg, it was the entire surviving
population, and the mutation matrix's devio substrate refused because of
it: applied=30, caught=29, holes=['devio_stop_drop'], OVERALL: FAIL.

gen_FXSAVE hands tcg_env and s->A0 to helper_fxsave, which performs the
whole area access itself, so the op stream shows a call and no qemu_st
for the reader to attribute.  The decode site holds the operand, and it
now says what the instruction does: a 512-byte store for FXSAVE, a
512-byte load for FXRSTOR, at the address the modrm names.  512 is
QEMU's own extent for it -- sizeof(X86LegacyXSaveArea), the size
helper_fxsave passes to access_prepare -- not a constant chosen here.

This is the second shape the synthetic-EA statement carries, and the
header said only the first.  A prefetch performs no access; RISC-V's
CBO.ZERO and x86's FXSAVE perform a real one inside a helper.  What they
share is the reader's problem -- no op to walk -- and the direction and
size are the helper's own reading in both.  The comment says so now
instead of asserting that the emulation computes no address.

The memop row's size field was a byte and is now 16 bits.  512 does not
fit in a byte, and the clamp would have published a size the instruction
does not touch through qemu_plugin_dataflow_memop::size, which is
already uint32_t.  No access in tree exceeded 255 before this one, so
the widening moves nothing else.

THE AREA'S REGISTERS ARE NOT NAMED, and the reason is a cap.  FXSAVE's
stored datum comes from the 32 registers the area holds and FXRSTOR
writes all of them; stating that was written, built and MEASURED here,
and REVERTED.  An instruction may carry INSN_DF_MAX_FIELDS (16) env
ranges and INSN_DF_MAX_WRITES (8) writes, the area names 32, and the
reader correctly refuses an instruction it cannot record in full -- both
encodings came back with src=[], dst=[] and no access at all.  So the
store's data provenance is empty where the area's registers belong, and
FXRSTOR still publishes no destination.  Both are named residues with
that blocker, not silent zeros.

MEASURED, same probe, same build dir:

    before  480fae4040  op=GEN_OP_STORE  (no deps line: no memop slot)
    after   480fae4040  op=GEN_OP_STORE  deps: sd=0x0 sa=0x3
    before  480fae4840  op=GEN_OP_LOAD   (no deps line)
    after   480fae4840  op=GEN_OP_LOAD   deps: la=0x3

and on the system devio trace the audit reads impossible=(0 memop, 0
regdata), strict rc=0, with the mutation matrix applied=30 caught=30
holes=[] OVERALL: PASS.

+++ include/exec/insn-dataflow.h
@@ -354,7 +354,13 @@ typedef struct InsnDataflowWrite {
  */
 typedef struct InsnDataflowMemop {
     uint8_t  dir;               /* INSN_DF_RD / _WR */
-    uint8_t  size;              /* bytes */
+    /*
+     * Bytes.  Wide enough for an area access an instruction performs inside a
+     * helper: x86's FXSAVE moves 512 bytes and a byte field would have had to
+     * clamp it, publishing a size the instruction does not touch through an
+     * API (qemu_plugin_dataflow_memop::size) that is already uint32_t.
+     */
+    uint16_t size;
     uint64_t addr_prov[INSN_DF_REG_WORDS];
     uint64_t data_prov[INSN_DF_REG_WORDS];
 } InsnDataflowMemop;
@@ -975,13 +981,26 @@ void insn_dataflow_note_vec_operand(uint32_t envofs, uint32_t bytes,
                                     unsigned dir);
 
 /*
- * An effective address the emulation computes no address for.
+ * An access the op stream does not carry, and the address it is made at.
  *
- * Prefetches and cache-maintenance operations -- x86 prefetch*, aarch64 PRFM
- * and DC CVAU, MIPS PREF and SYNCI -- lower to a NOP or to a bare block exit.
- * There is no memop, so there is nothing for the reader to walk, and a
- * consumer is handed an instruction that names an address in its encoding and
- * touches nothing.  The decode site still holds the operand, so it says so.
+ * Two shapes reach this, and the difference between them is whether the guest
+ * memory is touched -- not whether the reader can see it, which is the same
+ * answer for both: nothing in the op stream to walk.
+ *
+ * The first performs no access.  Prefetches and cache-maintenance operations
+ * -- x86 prefetch*, aarch64 PRFM and DC CVAU, MIPS PREF and SYNCI -- lower to
+ * a NOP or to a bare block exit, so a consumer is handed an instruction that
+ * names an address in its encoding and touches nothing.
+ *
+ * The second performs a REAL one, inside a helper.  RISC-V's CBO.ZERO memsets
+ * a cache block through a host pointer and x86's FXSAVE / FXRSTOR move the
+ * 512-byte area, and in both the helper takes tcg_env and an address and does
+ * the work itself: no qemu_ld/qemu_st op exists for the reader to attribute,
+ * so the template would say the instruction cannot touch memory while the
+ * runtime callbacks deliver the accesses.  @dir and @size are the helper's
+ * own reading of what it does, never a guess.
+ *
+ * The decode site holds the operand in both shapes, so it says so.
  *
  * @parts are the components the address is built from -- each an atom, the
  * shift the computation applies to it and any narrowing extend -- and @disp
